Fixed: fallback to OS temp directory if findCacheDir fails (#154)
* Fallback to os temp directory if findCacheDir fail

From the readme: `Finds the cache dir using the supplied options. The algorithm tries to find a package.json file, searching every parent directory of the cwd specified (or implied from other options). It returns a string containing the absolute path to the cache directory, or null if package.json was never found.` This is a fallback for null condition. Investigated after @gaearon comments on https://github.com/facebookincubator/create-react-app/pull/1578#discussion_r101802866

* remove semicolon to align with overall code style

* Use double quotes instead of single quotes

To pass CI